requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.0
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmake==4.0.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.10.0
debugpy==1.6.6
decorator==5.1.1
docker-pycreds==0.4.0
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.104.1
filelock==3.12.2
Flask==2.3.2
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.9.2
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
greenlet==3.0.1
grpcio==1.53.0
h11==0.14.0
h5py==3.8.0
haversine==2.8.0
huggingface-hub==0.16.4
idna==2.10
imageio==2.28.0
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
iterators==0.0.2
itsdangerous==2.1.2
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.1.0
jupyter_core==5.3.0
keras==2.10.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
lime==0.2.0.1
lit==18.1.8
llvmlite==0.39.1
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.0.4
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.6
networkx==3.1
ninja==1.11.1
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
openai==0.27.6
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==25.0
pandas==2.1.1
parso==0.8.3
pathtools==0.1.2
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pyclipper==1.3.0.post4
pydantic==2.4.2
pydantic_core==2.10.1
PyDirectInput==1.0.4
pygame==2.4.0
PyGetWindow==0.0.9
Pygments==2.14.0
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-version==0.0.2
python3-xlib==0.15
pytweening==1.0.4
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
regex==2023.10.3
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==14.0.0
roboflow==1.0.9
rsa==4.9
safetensors==0.3.3
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
seaborn==0.12.2
sentry-sdk==1.22.1
setproctitle==1.3.2
shap==0.41.0
shapely==2.0.1
six==1.15.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.23
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2023.4.12
tokenizers==0.14.0
torch==2.0.1
torchvision==0.15.2
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.34.0
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.124
urllib3==1.26.15
uvicorn==0.24.0.post1
wcwidth==0.2.6
Werkzeug==2.3.4
wget==3.2
wrapt==1.12.1
yarl==1.9.2
zipp==3.15.0
